Switched to C99 data types (fixing sound on 64-bits), removed the GX code, RS-97 port etc...

+++ source/sound/Fir_Resampler.h
@@ -4,26 +4,28 @@
 #ifndef FIR_RESAMPLER_H
 #define FIR_RESAMPLER_H
 
+#include <stdint.h>
+
 #define STEREO        2
 #define MAX_RES       32
 #define WIDTH         16
 #define WRITE_OFFSET  (WIDTH * STEREO) - STEREO
 #define GAIN          1.0
 
-typedef signed long int sample_t;
+typedef int32_t sample_t;
 
-extern int Fir_Resampler_initialize( int new_size );
+extern int32_t Fir_Resampler_initialize( int32_t new_size );
 extern void Fir_Resampler_shutdown( void );
 extern void Fir_Resampler_clear( void );
 extern double Fir_Resampler_time_ratio( double new_factor, double rolloff );
 extern double Fir_Resampler_ratio( void );
-extern int Fir_Resampler_max_write( void );
+extern int32_t Fir_Resampler_max_write( void );
 extern sample_t* Fir_Resampler_buffer( void );
-extern int Fir_Resampler_written( void );
-extern int Fir_Resampler_avail( void );
-extern void Fir_Resampler_write( long count );
-extern int Fir_Resampler_read( sample_t* out, long count );
-extern int Fir_Resampler_input_needed( long output_count );
-extern int Fir_Resampler_skip_input( long count );
+extern int32_t Fir_Resampler_written( void );
+extern int32_t Fir_Resampler_avail( void );
+extern void Fir_Resampler_write( int32_t count );
+extern int32_t Fir_Resampler_read( sample_t* out, int32_t count );
+extern int32_t Fir_Resampler_input_needed( int32_t output_count );
+extern int32_t Fir_Resampler_skip_input( int32_t count );
 
 #endif
